refactor: move fetching eedata to a hook
