+++ src/pages/EarthEngineImport/EarthEngineImportForm.js
@@ -16,21 +16,16 @@ import { OrganisationUnits } from './components/OrganisationUnits.js'
 import { Periods } from './components/Periods.js'
 import { Rounding, defaultRoundingOption } from './components/Rounding.js'
 import { Tooltip } from './components/Tooltip.js'
+import { useEeData } from './components/useEeData.js'
 import styles from './EarthEngineImportForm.module.css'
 import { onImport } from './form-helper.js'
-import { getPeriods, getAggregations } from './util/earthEngineHelper.js'
-import {
-    POPULATION_AGE_GROUPS_DATASET_ID,
-    getDefaultAggregation,
-} from './util/earthEngines.js'
+import { POPULATION_AGE_GROUPS_DATASET_ID } from './util/earthEngines.js'
 import {
     ORGANISATION_UNITS,
     ROUNDING,
     DATA_ELEMENT_ID,
     EARTH_ENGINE_ID,
 } from './util/formFieldConstants.js'
-import { getEarthEngineConfig } from './util/getEarthEngineConfig.js'
-import { getStructuredData } from './util/getStructuredData.js'
 
 const { Form, Field, FormSpy } = ReactFinalForm
 
@@ -55,11 +50,9 @@ const EarthEngineImportForm = () => {
 
     const [progress, setProgress] = useState(false)
     const [showFullSummaryTask, setShowFullSummaryTask] = useState(false)
-    const [eeData, setEeData] = useState([])
-    const [pointOuRows, setPointOuRows] = useState(null)
-    const [fetching, setFetching] = useState(false)
-    const [doSubmit, setDoSubmit] = useState(false)
-    const [requestFailedMessage, setRequestFailedMessage] = useState(null)
+
+    const { eeData, pointOuRows, loading, errorMessage, fetchEeData } =
+        useEeData()
 
     // for scrolling to the Job summary which is placed at the top
     const hiddenTopElRef = useRef(null)
@@ -70,86 +63,6 @@ const EarthEngineImportForm = () => {
         [DATA_ELEMENT_ID]: null,
     }
 
-    const fetchEeData = async ({
-        earthEngineId,
-        organisationUnits,
-        associatedGeometry,
-        period,
-        rounding,
-        bandCocs = [],
-    }) => {
-        const selectedBandCocs = bandCocs.filter((bc) => !!bc.coc)
-
-        const aggregationType = getDefaultAggregation(earthEngineId)
-
-        setDoSubmit(false)
-        setFetching(true)
-        setRequestFailedMessage(null)
-
-        try {
-            const periods = await getPeriods(earthEngineId, engine)
-
-            const { config, pointOrgUnits } = await getEarthEngineConfig(
-                {
-                    earthEngineId,
-                    organisationUnits,
-                    associatedGeometry,
-                    period,
-                    periods,
-                    aggregationType,
-                    selectedBandCocs,
-                },
-                engine
-            )
-
-            const data = await getAggregations(engine, config)
-
-            const ouIdNameMap = config.data?.reduce((acc, curr) => {
-                acc[curr.id] = curr.properties
-                return acc
-            }, {})
-
-            const structuredData = getStructuredData({
-                data,
-                selectedBandCocs,
-                ouIdNameMap,
-                rounding,
-                aggregationType,
-            })
-
-            setEeData(structuredData)
-            setPointOuRows(pointOrgUnits)
-            setFetching(false)
-            setDoSubmit(true)
-        } catch (error) {
-            console.log('Error while fetching Earth Engine data', error)
-            setEeData([])
-            setFetching(false)
-            const message = error.message || error
-
-            let msg = i18n.t(
-                'An error occurred while trying to fetch Earth Engine data'
-            )
-
-            if (message.includes('Output of image computation is too large')) {
-                msg = i18n.t(
-                    'The Earth Engine data set is too large. Try reducing the number of groups or organisation units'
-                )
-            } else if (
-                message.includes(
-                    'Dimension is present in query without any valid dimension options: `ou`'
-                )
-            ) {
-                msg = i18n.t('The organisation units selection is invalid')
-            } else if (message.length) {
-                msg = msg.concat(`: ${message}`)
-            }
-            setRequestFailedMessage(msg)
-
-            return
-        }
-    }
-
     const previewIsAllowed = ({ valid, values, modifiedSinceLastSubmit }) => {
         // there should be at least one band for Population Age groups
         const bandsValid =
@@ -157,9 +70,9 @@ const EarthEngineImportForm = () => {
                 ? !!values.bandCocs?.find((bc) => !!bc.coc)
                 : true
 
-        const otherCheck = requestFailedMessage ? modifiedSinceLastSubmit : true
+        const modified = errorMessage ? modifiedSinceLastSubmit : true
 
-        return valid && bandsValid && otherCheck
+        return valid && bandsValid && modified
     }
 
     const wrappedSetShowFullSummaryTask = (val) => {
@@ -175,17 +88,17 @@ const EarthEngineImportForm = () => {
     })
 
     const onSubmit = (values) => {
-        if (doSubmit) {
+        if (!loading && !errorMessage) {
             onImportInternal({ eeData, ...values })
         }
     }
 
     const getAlerts = (submitError) => {
         const internalErrors = []
-        if (requestFailedMessage) {
+        if (errorMessage) {
             internalErrors.push({
                 id: 'getEeDataFailed',
-                message: requestFailedMessage,
+                message: errorMessage,
                 critical: true,
                 info: false,
                 warning: false,
@@ -295,7 +208,7 @@ const EarthEngineImportForm = () => {
                                         {({ modifiedSinceLastSubmit }) =>
                                             !modifiedSinceLastSubmit ? (
                                                 <DataPreview
-                                                    fetching={fetching}
+                                                    loading={loading}
                                                     eeData={eeData}
                                                     pointOuRows={pointOuRows}
                                                 />
@@ -309,7 +222,7 @@ const EarthEngineImportForm = () => {
                                     >
                                         {({ modifiedSinceLastSubmit }) =>
                                             !modifiedSinceLastSubmit &&
-                                            !fetching &&
+                                            !loading &&
                                             eeData?.length ? (
                                                 <div
                                                     className={
